requirements.txt
Flask==3.0.2
selenium==4.18.1
beautifulsoup4==4.12.3
google-generativeai==0.3.2
python-dotenv==1.0.1
pytest==8.1.1
webdriver-manager==4.0.1 